We can't assume a story has authors when trying to build the story label.
This bug was hidden before the migration was dropping all the
departments and story types we don't care about for 3.0.  Now that those
terms are gone, it's clear there are many more cases where the logic
to build a story label has no matches and we have to leave it blank.