Migrate: Node migrations should always use nid, not tnid, to avoid possible weirdness.
In some of the migration runs, I saw hicups with node migrations where
the source and destination nid didn't match. I suspected it was from
the translation-related 'tnid' stuff, which apparently isn't fully
stable. Switching to rely on the original 'nid' gets everything
working solid, and since we don't use any of the multi-lingual support
(yet), it's safe to ignore this complication.