[SW #983] Migrate: Follow-up fixes for dropping terms without nodes.
There seems to be a bug with SkipOnEmpty that skips the whole row when
a process is skipped in some circumstances. Haven't gotten to the
bottom of it, but it was causing us major trouble when trying to skip
terms without nodes (see commit 0acb4644). The initial test for our
'valid_term' property worked, but then if we were trying to migrate a
term before importing its parent, the whole row would be skipped.

Since a) we're preserving all IDs (including tids) during migration,
and b) we're only pruning terms from vocabularies that don't have
heirarchies, it's safe to remove the code that's trying to lookup the
appropriate parent ID and skip when appropriate. We can always just
migrate parent directly from the D6 site and it all Just Works(tm).