Story: Don't assume 'Interview' stories have field_interviewees defined.
For example, /node/37 is an interview with an anonymous immigrants
rights activist. So, default the story label to "Interview" and only
append the ': ' and names if field_interviewees has a value.

Ditto 'Obituary'. (I guess. I'm not sure this happens, but it doesn't
hurt to be careful and handle the case, anyway).