[SW #877]: Add a custom KernelEvents::REQUEST event listener to redirect *.shml to *.php:
We don't want to do this directly in setting.php since we want to give
redirect.module a chance to redirect specific legacy .shtml paths to
the Drupal nodes (when possible). Core's listener for this event uses
priority 32. Redirect uses 33.  We need to happen after redirect but
before core, so we also use priority 33 but set our module weight
heavier than redirect.