[SW #990]: Further reading: Add static caching to avoid rebuilding the block:
Whenever the cache is cleared, an article is edited, etc, the next
page load needs to build this block twice (sidebar vs. story footer).
Sadly, the render cache doesn't save us, and by default we'd re-do all
the queries, node loads, etc. This adds a static cache to the
SWFurtherReadingBlock class so once we find + render everything, we
save the render array, and can immediately re-use it for the 2nd block.