[SW #1046]: Topic filter on 'Find stories' searches main or secondary topic:
Instead of using direct field filters on the entity (term) reference fields,
we can use the magic "Has taxonomy term" (taxonomy_index_tid) filter, limit
it to the Topic vocabulary, and it all works as expected. There are records
in the {taxonomy_index} table for every term reference, regardless of what
field it's from. Tid 177 (Labor) looks the same, whether main or secondary.