[SW #1039]: Styles for cover image on print edition pages.
The changes from commit bcf77cb5112 related to this broke the display
of images with the top stories field, so re-doing this to be more
specific in this case to only target the cover image.

Also adding top margin to the top stories field so those don't crash
into the cover image or description.